+++ ewm-service/src/main/java/ru/practicum/category/service/CategoryPublicServiceBase.java
@@ -7,7 +7,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import ru.practicum.category.repository.CategoryRepository;
 import ru.practicum.dto.category.CategoryDto;
-import ru.practicum.dto.category.CategoryMapper;
+import ru.practicum.util.mapper.CategoryMapper;
 import ru.practicum.model.Category;
 import ru.practicum.util.exception.NotFoundException;
