+++ ewm-service/src/main/java/ru/practicum/compilation/service/CompilationAdminServiceBase.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.practicum.compilation.repository.CompilationRepository;
 import ru.practicum.dto.compilation.CompilationDto;
-import ru.practicum.dto.compilation.CompilationMapper;
+import ru.practicum.util.mapper.CompilationMapper;
 import ru.practicum.dto.compilation.NewCompilationDto;
 import ru.practicum.dto.compilation.UpdateCompilationRequest;
 import ru.practicum.event.repository.EventRepository;
