+++ ewm-service/src/main/java/ru/practicum/event/service/EventAdminServiceBase.java
@@ -8,11 +8,11 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import ru.practicum.category.repository.CategoryRepository;
 import ru.practicum.dto.event.EventFullDto;
-import ru.practicum.dto.event.EventMapper;
+import ru.practicum.util.mapper.EventMapper;
 import ru.practicum.dto.event.UpdateEventAdminRequest;
 import ru.practicum.dto.event.enums.StateActionAdmin;
 import ru.practicum.dto.location.LocationDto;
-import ru.practicum.dto.location.LocationMapper;
+import ru.practicum.util.mapper.LocationMapper;
 import ru.practicum.event.repository.EventRepository;
 import ru.practicum.event.repository.LocationRepository;
 import ru.practicum.model.Category;
