+++ ewm-service/src/main/java/ru/practicum/event/service/EventPrivateServiceBase.java
@@ -7,7 +7,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import ru.practicum.category.repository.CategoryRepository;
 import ru.practicum.dto.event.EventFullDto;
-import ru.practicum.dto.event.EventMapper;
+import ru.practicum.util.mapper.EventMapper;
 import ru.practicum.dto.event.EventShortDto;
 import ru.practicum.dto.event.NewEventDto;
 import ru.practicum.dto.event.UpdateEventUserRequest;
@@ -16,9 +16,9 @@ import ru.practicum.dto.event.enums.StateActionUser;
 import ru.practicum.dto.event.request.EventRequestStatusUpdateRequest;
 import ru.practicum.dto.event.request.EventRequestStatusUpdateResult;
 import ru.practicum.dto.event.request.ParticipationRequestDto;
-import ru.practicum.dto.event.request.ParticipationRequestMapper;
+import ru.practicum.util.mapper.ParticipationRequestMapper;
 import ru.practicum.dto.location.LocationDto;
-import ru.practicum.dto.location.LocationMapper;
+import ru.practicum.util.mapper.LocationMapper;
 import ru.practicum.event.repository.EventRepository;
 import ru.practicum.event.repository.LocationRepository;
 import ru.practicum.model.Category;
