+++ ewm-service/src/main/java/ru/practicum/user/service/UserServiceBase.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.dto.user.NewUserRequest;
 import ru.practicum.dto.user.UserDto;
-import ru.practicum.dto.user.UserMapper;
+import ru.practicum.util.mapper.UserMapper;
 import ru.practicum.model.User;
 import ru.practicum.user.repository.UserRepository;
 import ru.practicum.util.exception.NotFoundException;
