+++ ewm-service/src/main/java/ru/practicum/util/mapper/CompilationMapper.java
@@ -1,4 +1,4 @@
-package ru.practicum.dto.compilation;
+package ru.practicum.util.mapper;
 
 import org.mapstruct.BeanMapping;
 import org.mapstruct.Mapper;
@@ -6,7 +6,9 @@ import org.mapstruct.Mapping;
 import org.mapstruct.MappingConstants;
 import org.mapstruct.MappingTarget;
 import org.mapstruct.NullValuePropertyMappingStrategy;
-import ru.practicum.dto.event.EventMapper;
+import ru.practicum.dto.compilation.CompilationDto;
+import ru.practicum.dto.compilation.NewCompilationDto;
+import ru.practicum.dto.compilation.UpdateCompilationRequest;
 import ru.practicum.model.Compilation;
 
 @Mapper(componentModel = MappingConstants.ComponentModel.SPRING, uses = {EventMapper.class})
