+++ ewm-service/src/main/java/ru/practicum/util/mapper/UserMapper.java
@@ -1,8 +1,11 @@
-package ru.practicum.dto.user;
+package ru.practicum.util.mapper;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.MappingConstants;
+import ru.practicum.dto.user.NewUserRequest;
+import ru.practicum.dto.user.UserDto;
+import ru.practicum.dto.user.UserShortDto;
 import ru.practicum.model.User;
 
 @Mapper(componentModel = MappingConstants.ComponentModel.SPRING)
